# server/swagger-auto-generated.yaml
openapi: 3.0.2
info:
  title: Campus Events and Entertainment Center API
  version: 1.0.0
  description: API for managing campus events and entertainment
servers:
  - url: http://localhost:5000/api
  - url: http://217.142.229.202/api
  - url: http://10.27.41.93:5000/api
security:
  - bearerAuth: []
paths:
...
  /images:
    post:
      tags:
        - Images
      summary: Upload an image
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required:
                - image
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Image uploaded successfully
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - url
                properties:
                  url:
                    type: string
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /images/{image_id}:
    get:
      tags:
        - Images
      summary: Get an image by id
      parameters:
        - name: image_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Image found
          content:
            image/*:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/404'
  /users:
    post:
      tags:
        - Users
      summary: Register a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - username
                - password
                - user_email
              properties:
                username:
                  type: string
                  pattern: ^[a-zA-Z0-9]+$
                  maxLength: 15
                password:
                  type: string
                  pattern: ^[a-zA-Z0-9]+$
                  maxLength: 32
                user_email:
                  type: string
                  format: email
      responses:
        '200':
          description: User created successfully
        '429':
          $ref: '#/components/responses/429'
      security: []
  /sessions:
    post:
      tags:
        - Users
      summary: Login a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  pattern: ^[a-zA-Z0-9]+$
                  maxLength: 15
                password:
                  type: string
                  pattern: ^[a-zA-Z0-9]+$
                  maxLength: 32
      responses:
        '200':
          description: User logged in successfully
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  token:
                    type: string
                  expires:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/401'
      security: []
    delete:
      tags:
        - Users
      summary: Logout a user
      responses:
        '200':
          description: User logged out successfully
        '401':
          $ref: '#/components/responses/401'
  /users/{username}:
    get:
      tags:
        - Users
      summary: Get a user by username
      parameters:
        - $ref: '#/components/parameters/path_username'
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/404'
  /me:
    get:
      tags:
        - Users
      summary: Get the current user by token
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/401'
    put:
      tags:
        - Users
      summary: Edit the current user's info
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User info updated successfully
        '401':
          $ref: '#/components/responses/401'
    delete:
      tags:
        - Users
      summary: Delete the current user
      responses:
        '200':
          description: User deleted successfully
        '401':
          $ref: '#/components/responses/401'
  /messages:
    get:
      tags:
        - Users
      summary: Get messages sent to the current user
      responses:
        '200':
          description: Messages found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/401'
components:
  parameters:
    path_username:
      name: username
      in: path
      required: true
      schema:
        type: string
        pattern: ^[a-zA-Z0-9]+$
        maxLength: 15
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Event:
      type: object
      additionalProperties: false
      required:
        - id
        - title
        - description
        - poster
        - organizer
        - publish_organization
        - participants
        - time
        - location
        - tags
        - status
        - comments
        - ratings
      properties:
        id:
          type: integer
          readOnly: true
        title:
          type: string
        description:
          type: string
        poster:
          type: string
          description: The URL of the event's poster
        organizer_id:
          type: integer
        publish_organization:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        status:
          type: integer
        location:
          type: string
        capacity:
          type: integer
...
    Gender:
      enum:
        - 1
        - 2
        - 3
      type: integer
      description: 1 = male, 2 = female, 3 = other
    Message:
      type: object
      additionalProperties: false
      required:
        - title
        - content
        - time
        - sender
        - receiver
      properties:
        title:
          type: string
        content:
          type: string
        time:
          type: string
          format: date-time
        sender:
          $ref: '#/components/schemas/User'
        receiver:
          $ref: '#/components/schemas/User'
    ErrorDetail:
      title: ErrorDetail
      additionalProperties: false
      required:
        - title
        - description
      type: object
      properties:
        title:
          title: Title
          type: string
        description:
          title: Description
          type: string
    User:
      type: object
      additionalProperties: false
      required:
        - username
        - nickname
        - user_group
        - avatar
        - user_intro
        - user_email
        - gender
        - birthday
        - event_history
        - following
        - followers
        - published_events
      properties:
        id:
          type: integer
          description: The auto-generated id of the user
          readOnly: true
        username:
          type: string
          description: The username of the user
        nickname:
          type: string
        user_group:
          $ref: '#/components/schemas/UserGroup'
        avatar:
          type: string
          description: The URL of the user's avatar
        user_intro:
          type: string
        user_email:
          type: string
          format: email
        password:
          type: string
          description: The password after sha256 encryption with salt
        gender:
          $ref: '#/components/schemas/Gender'
        birthday:
          type: string
          format: date
        event_history:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        following:
          type: array
          items:
            $ref: '#/components/schemas/User'
        followers:
          type: array
          items:
            $ref: '#/components/schemas/User'
        published_events:
          type: array
          items:
            $ref: '#/components/schemas/Event'
    UserGroup:
      enum:
        - 1
        - 2
        - 3
      type: integer
      description: 1 = Audience, 2 = Organizer, 3 = Admin
  responses:
    '400':
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
    '401':
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
...
    '404':
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
    '429':
      description: Too Many Requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
    '500':
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
...